test-requirements.txt
django-wysiwyg
html5lib==1.0b10
Pillow
micawber
docutils
textile
MarkDown==2.6.8
pygments
disqus
#twitter-text-py
